The data files install as package_data instead of data_files.
This seams cleaner than hacking with INSTALL_SCHEMES, which didn't even work in my environment (Python 3.6, Windows, virtualenv, pip install) and the data files ended up in the virtualenv root instead of the package dir.

Note: the fixup was not tested in any other environment, hopefully it will not break anything.